+++ R/get_ag_bulletin.R
@@ -51,10 +51,10 @@
 #' @export
 get_ag_bulletin <- function(state = "AUS") {
   # CRAN NOTE avoidance
-  stations_site_list <- NULL
+  stations_site_list <- NULL # nocov
 
   # Load AAC code/town name list to join with final output
-  load(system.file("extdata", "stations_site_list.rda", package = "bomrang"))
+  load(system.file("extdata", "stations_site_list.rda", package = "bomrang")) # nocov
 
   the_state <- .check_states(state) # see internal_functions.R